Generalized some core data structs
