+++ lib/hashcraft/compiler.rb
@@ -7,8 +7,6 @@
 # LICENSE file in the root directory of this source tree.
 #
 
-require_relative 'dsl'
-
 module Hashcraft
   # This class understands how to traverse an option chain and output a hash.
   class Compiler
