+++ lib/hashcraft/generic.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+#
+# Copyright (c) 2020-present, Blue Marble Payroll, LLC
+#
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+#
+
+require_relative 'generic/dictionary'
+require_relative 'generic/registry'
