+++ lib/hashcraft/transformer_registry.rb
@@ -7,14 +7,13 @@
 # LICENSE file in the root directory of this source tree.
 #
 
-require_relative 'registry'
 require_relative 'transformers/camel_case'
 require_relative 'transformers/pascal_case'
 require_relative 'transformers/pass_thru'
 
 module Hashcraft
   # Singleton that knows how to register and retrieve transformer instances.
-  class TransformerRegistry < Registry
+  class TransformerRegistry < Generic::Registry
     def initialize
       super(
         'camel_case' => Transformers::CamelCase.instance,
